improving transaction input of currency

+++ app/src/main/java/com/aleexalvz/cashwise/data/model/investment/Investment.kt
@@ -5,15 +5,15 @@ data class Investment(
     val userID: Long,
     val title: String,
     val category: InvestmentCategory,
-    val unitValue: Double,
+    val unitValue: Long,
     val amount: Long,
     val type: InvestmentType,
     val dateMillis: Long
 )
 
-fun Investment.totalValue(): Double = amount * unitValue
+fun Investment.totalValue(): Long = amount * unitValue
 
-fun List<Investment>.totalValue(): Double = fold(0.0) { accumulator, investment ->
+fun List<Investment>.totalValue(): Long = fold(0) { accumulator, investment ->
     when (investment.type) {
         InvestmentType.PROFIT -> (accumulator + investment.totalValue())
         InvestmentType.LOSS -> (accumulator - investment.totalValue())
